requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
motor==3.4.0
pydantic==1.10.11
pymongo==4.6.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
aiosmtplib==1.1.5
google-auth==2.22.0
httpx==0.24.1
loguru==0.7.0
slowapi==0.1.8